requirements.txt
# Core dependencies
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0

# WordPress API
python-wordpress-xmlrpc>=2.3

# Social Media APIs
facebook-sdk>=3.1.0

# Scheduling
schedule>=1.2.0
APScheduler>=3.10.0

# Data processing
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# Database
sqlite3

# Logging
colorlog>=6.7.0
